Refactor build_op_downstream_map for standalone executor
